Move content type reindexing to its own service
