+++ src/Umbraco.Cms.Search.Core/Services/ContentIndexing/IContentIndexingService.cs
@@ -1,4 +1,3 @@
-using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Search.Core.Models.Indexing;
 
 namespace Umbraco.Cms.Search.Core.Services.ContentIndexing;
@@ -8,6 +7,4 @@ public interface IContentIndexingService
     void Handle(IEnumerable<ContentChange> changes, string origin);
 
     void Rebuild(string indexAlias, string origin);
-
-    void ReindexByContentTypes(Guid[] contentTypeKeys, UmbracoObjectTypes objectType, string origin);
 }
