+++ src/Umbraco.Cms.Search.Core/Services/ContentIndexing/IContentTypeIndexingService.cs
@@ -1,0 +1,8 @@
+﻿using Umbraco.Cms.Core.Models;
+
+namespace Umbraco.Cms.Search.Core.Services.ContentIndexing;
+
+public interface IContentTypeIndexingService
+{
+    void ReindexByContentTypes(Guid[] contentTypeKeys, UmbracoObjectTypes objectType, string origin);
+}
